+++ apps/frontend/src/app/2_molecules/AmountRenderer/AmountRenderer.tsx
@@ -19,6 +19,7 @@ import {
   formatValue,
   getDecimalPartLength,
   getLocaleSeparators,
+  isScientificNumber,
 } from '../../../utils/math';
 import { BigNumber } from 'ethers';