+++ apps/frontend/src/app/5_pages/EarnPage/EarnPage.tsx
@@ -31,6 +31,7 @@ import {
   TabType,
 } from '@sovryn/ui';
 
+import { LedgerPermitLocked } from '../../1_atoms/LedgerPermitLocked/LedgerPermitLocked';
 import { AmountRenderer } from '../../2_molecules/AmountRenderer/AmountRenderer';
 import { AssetRenderer } from '../../2_molecules/AssetRenderer/AssetRenderer';
 import { MaxButton } from '../../2_molecules/MaxButton/MaxButton';
@@ -41,6 +42,7 @@ import { useAssetBalance } from '../../../hooks/useAssetBalance';
 import { useBlockNumber } from '../../../hooks/useBlockNumber';
 import { useMaintenance } from '../../../hooks/useMaintenance';
 import { translations } from '../../../locales/i18n';
+import { LEDGER } from '../../../utils/constants';
 import {
   decimalToBn,
   formatValue,
@@ -63,7 +65,7 @@ const EarnPage: FC = () => {
   const [token, setToken] = useState<SupportedTokens>(SupportedTokens.dllr);
   const [isLoading, setIsLoading] = useState(false);
 
-  const { account } = useAccount();
+  const { account, type } = useAccount();
   const { value: block } = useBlockNumber();
 
   const { liquity } = useLoaderData() as {
@@ -296,6 +298,11 @@ const EarnPage: FC = () => {
     [],
   );
 
+  const ledgerAndDllr = useMemo(
+    () => type === LEDGER && token === SupportedTokens.dllr && isDeposit,
+    [isDeposit, token, type],
+  );
+
   return (
     <>
       <Helmet>
@@ -363,6 +370,9 @@ const EarnPage: FC = () => {
               dataAttribute="earn-amount-input-error"
             />
           )}
+
+          {ledgerAndDllr && <LedgerPermitLocked />}
+
           <SimpleTable className="mt-3">
             <SimpleTableRow
               label={t(pageTranslations.currentPoolBalance)}
@@ -401,7 +411,7 @@ const EarnPage: FC = () => {
             text={t(commonTranslations.buttons.confirm)}
             className="w-full mt-8"
             onClick={handleSubmit}
-            disabled={isSubmitDisabled}
+            disabled={isSubmitDisabled || ledgerAndDllr}
             dataAttribute="earn-submit"
           />
           {isInMaintenance && (